% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{oracl_list_to_df}
\alias{oracl_list_to_df}
\title{Add list names as column and combine}
\usage{
oracl_list_to_df(als)
}
\arguments{
\item{als}{A list of ORA dataframes}
}
\value{
One ORA dataframe
}
\description{
Takes a list of ORAs, adds column "grouping" with element names and combines to a single ORA dataframe
}
