% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{oraclient}
\alias{oraclient}
\title{Panther GO term enrichment}
\usage{
oraclient(
  x,
  bg = NULL,
  ontology = "bp",
  taxon = "Athaliana",
  enrichmentTestType = "FISHER",
  correction = "FDR",
  panther_api.url = "http://pantherdb.org//services/oai/pantherdb/enrich/overrep",
  fdr.thresh = 0.1,
  p.thresh = 0.05
)
}
\arguments{
\item{x}{A list of gene identifiers (character vector)}

\item{bg}{A list of background genes (character vector). If not provided, uses PANTHER standard background.}

\item{ontology}{Panther annotation data type of interest (Defaults to "GO:0008150" for BP, also possible: "GO:0003674" for MF, "GO:0005575" for CC)}

\item{taxon}{Name of organism to be studied (Defaults to: Athaliana (ID: 3702))}

\item{enrichmentTestType}{Type of enrichment test (Defaults to: "FISHER", also possible: BINOMIAL)}

\item{correction}{FP-Correction method (Defaults to: "FDR", also possible: BONFERRONI)}

\item{panther_api.url}{URL to API "http://pantherdb.org//services/oai/pantherdb/enrich/overrep"}

\item{fdr.thresh}{FDR cutoff (default: 0.1)}

\item{p.thresh}{P-value cutoff to be applied (default: p < 0.05)}
}
\value{
A dataframe or panther output
}
\description{
A client for the panther webservice. Takes a list of genes and a background geneset and sends them to panther for a classic Overrepresentation Analysis (ORA)
}
