% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{oraclient_add_genes}
\alias{oraclient_add_genes}
\title{Add Gene Identifiers}
\usage{
oraclient_add_genes(df, genes, ont)
}
\arguments{
\item{als}{Each and every entry in results$result of a Panther Enrichment}
}
\value{
Tabular GOterms to be processed further
}
\description{
Add Locus IDs for overrepresented GOterms
}
