% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{oraclient_body}
\alias{oraclient_body}
\title{Translate ontology and taxon to IDs}
\usage{
oraclient_body(geneset, geneset_bg = NULL, ont, tax, ett, correct)
}
\arguments{
\item{geneset}{User-specified geneset}

\item{geneset_bg}{Background. If null, use Panther standard}

\item{ont}{ontology to be investigated (ID!)}

\item{tax}{User-specified taxon, must be "Athaliana".}

\item{ett}{enrichmentTestType}

\item{correct}{correction method}
}
\value{
A vector of length 2 with Panther-specific IDs for taxon and ontology.
}
\description{
translate ontology and taxon to IDs readable by the panther API.
}
