% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{oraclient_extract_goterms}
\alias{oraclient_extract_goterms}
\title{Extract GOterms from Panther output}
\usage{
oraclient_extract_goterms(als)
}
\arguments{
\item{als}{Each and every entry in results$result of a Panther Enrichment}
}
\value{
Tabular GOterms to be processed further
}
\description{
Extracts GOterms from Panther response
}
