% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{oraclient_POST}
\alias{oraclient_POST}
\title{POST request}
\usage{
oraclient_POST(post_body, post_url)
}
\arguments{
\item{post_body}{Body to be used (a list)}

\item{post_url}{URL to be used (character)}
}
\value{
An ORA (.json)
}
\description{
Use predefined body and API-URL for post request
}
