% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{oraclient_settings}
\alias{oraclient_settings}
\title{Translate ontology and taxon to IDs}
\usage{
oraclient_settings(ont, tax)
}
\arguments{
\item{ont}{User-specified ontology, one of "bp", "mf" or "cc".}

\item{tax}{User-specified taxon, must be "Athaliana".}
}
\value{
A vector of length 2 with Panther-specific IDs for taxon and ontology.
}
\description{
translate ontology and taxon to IDs readable by the panther API.
}
