% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{oraclient_to_df}
\alias{oraclient_to_df}
\title{Oraclient to dataframe}
\usage{
oraclient_to_df(oraclient.ls)
}
\arguments{
\item{oraclient.ls}{A list of goterms}
}
\value{
An ORA dataframe (tibble)
}
\description{
Build a dataframe from output of oracl::oraclient_POST()
}
