% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{oraclot}
\alias{oraclot}
\title{Dot Plot}
\usage{
oraclot(df, top_n = 50)
}
\arguments{
\item{df}{A dataframe generated using `oracl::oraclient`.}

\item{top_n}{if set, number of terms that should be labelled. Default: `50`. Select `NULL` for no filter at all (not recommended, will get cluttered).}
}
\description{
Creates a dot plot from the output of `oracl::oraclient`.
}
