% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcanoracl}
\alias{volcanoracl}
\title{Volcano Plot}
\usage{
volcanoracl(df, x = "fold_change", y = "fdr", top_n = 5)
}
\arguments{
\item{df}{A dataframe generated using `oracl::oraclient`}

\item{x}{NUMERIC column to be used for the x axis of the plot. Will be log2-transformed. (default: `"fold_change"`)}

\item{y}{Significance metric to be used for the x axis of the plot. Will be -log-transformed. Options: `c("p.value", "fdr")` (default: `"fdr"`)}

\item{top_n}{Number of terms that should be labelled. Default: 5.}
}
\description{
Creates a volcano plot from the output of `oracl::oraclient`
}
