% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{oraclot}
\alias{oraclot}
\title{Dot Plot}
\usage{
oraclot(df, top_n = 50, trunc = 20, side = "left", include_ID = TRUE)
}
\arguments{
\item{df}{A dataframe generated using `oracl::oraclient`.}

\item{top_n}{if set, number of terms that should be labelled. Default: `50`. Select `NULL` for no filter at all (not recommended, will get cluttered).}

\item{trunc}{To which size should Term-label be truncated? Defaults to `20`}

\item{side}{From which side should truncation happen? Defauls to `left`}

\item{include_ID}{should GO_ID be included in plot labels? Defaults to `TRUE`}
}
\description{
Creates a dot plot from the output of `oracl::oraclient`.
}
